refactor: remove context and use url pram to handle sorting
