+++ src/app/page.tsx
@@ -6,18 +6,14 @@ import Image from 'next/image'
 import HotelsListing from '@/components/HotelsListing'
 import PageWrapper from '@/components/PageWrapper'
 
-import { HotelsProvider } from '@/context'
-
 const Page: FC = () => (
-  <HotelsProvider>
-    <main>
-      <PageWrapper>
-        <Image src="/images/qantas-logo.png" alt="Qantas Logo" width={145} height={28} priority />
+  <main>
+    <PageWrapper>
+      <Image src="/images/qantas-logo.png" alt="Qantas Logo" width={145} height={28} priority />
 
-        <HotelsListing />
-      </PageWrapper>
-    </main>
-  </HotelsProvider>
+      <HotelsListing />
+    </PageWrapper>
+  </main>
 )
 
 export default Page
