+++ src/components/ListingResult/ListingResult.tsx
@@ -1,14 +1,19 @@
+import { useRouter, useSearchParams } from 'next/navigation'
 import Sort from '../Sort'
 import { Wrapper } from './ListingResult.styles'
 
-import useHotels from '@/hooks/useHotels'
+import { SortEnum } from '@/types'
 
 type ListingResultProps = {
   numberOfItems: number
 }
 
 const ListingResult = ({ numberOfItems = 0 }: ListingResultProps) => {
-  const { sortBy } = useHotels()
+  const router = useRouter()
+  const params = useSearchParams()
+
+  const sortOrder = params.get('sort') || ''
+  const handleSort = (sort: SortEnum) => router.push(`?sort=${sort}`)
 
   return (
     <Wrapper>
@@ -16,7 +21,7 @@ const ListingResult = ({ numberOfItems = 0 }: ListingResultProps) => {
         <b data-testid="hotel-counts">{numberOfItems}</b> <i>hotels in</i> <b>Sydney</b>
       </p>
 
-      <Sort onChange={sortBy} />
+      <Sort onChange={handleSort} sortOrder={sortOrder as SortEnum} />
     </Wrapper>
   )
 }
