--- src/context/index.ts
@@ -1,3 +1,0 @@
-export { default as HotelsContext } from './HotelsContext'
-export { default as HotelsProvider } from './HotelsProvider'
-export * from './HotelsContext'
