--- src/hooks/useHotels/useHotels.tsx
@@ -1,8 +1,0 @@
-import { useContext } from 'react'
-
-import type { Context } from '@/context/HotelsContext'
-import HotelsContext from '@/context/HotelsContext'
-
-const useHotels = (): Context => useContext(HotelsContext)
-
-export default useHotels
